feat(2): rework header of file info

+++ src/BitMaps/BitMapFilesInfo.h
@@ -9,7 +9,7 @@ struct BitMapFilesInfo{
     uint8_t nameOfFile[20];
     uint32_t sizeOfFile;
     uint16_t sizeOfAlphabet;
-    uint8_t * symbols;
+    uint16_t * symbols;
     uint32_t * countOfThatSymbol;
 };
 
@@ -30,7 +30,7 @@ void WriteBitMapFileInfo(const struct BitMapFilesInfo* bitMapFilesInfo,FILE* fil
           1,
           file);
     fwrite(bitMapFilesInfo->symbols,
-           sizeof(uint8_t) * bitMapFilesInfo->sizeOfAlphabet,
+           sizeof(uint16_t) * bitMapFilesInfo->sizeOfAlphabet,
            1,
            file);
     fwrite(bitMapFilesInfo->countOfThatSymbol,
@@ -54,10 +54,10 @@ void ReadBitMapFileInfo(struct BitMapFilesInfo* bitMapFilesInfo,FILE* file,const
           sizeof(uint16_t),
           1,
           file);
-    bitMapFilesInfo->symbols = (uint8_t*) malloc(sizeof(uint8_t)*bitMapFilesInfo->sizeOfAlphabet);
+    bitMapFilesInfo->symbols = (uint16_t*) malloc(sizeof(uint16_t)*bitMapFilesInfo->sizeOfAlphabet);
     bitMapFilesInfo->countOfThatSymbol = (uint32_t*) malloc(sizeof(uint32_t)*bitMapFilesInfo->sizeOfAlphabet);
     fread(bitMapFilesInfo->symbols,
-           sizeof(uint8_t) * bitMapFilesInfo->sizeOfAlphabet,
+           sizeof(uint16_t) * bitMapFilesInfo->sizeOfAlphabet,
            1,
            file);
     fread(bitMapFilesInfo->countOfThatSymbol,
